move value objects to entities
