+++ src/contexts/CoreContext/domain/aggregates/Freelancer.ts
@@ -1,10 +1,10 @@
 import { UniqueEntityID } from "@/contexts/Shared/domain/UniqueEntityID";
 import { About } from "../valueObjects/About";
-import { Language } from "../valueObjects/Language";
-import { Skill } from "../valueObjects/Skill";
+import { Language } from "../entities/Language";
+import { Skill } from "../entities/Skill";
 import { Education } from "../entities/Education";
 import { Experience } from "../entities/Experience";
-import { Certification } from "../valueObjects/Certification";
+import { Certification } from "../entities/Certification";
 import { UserId } from "../valueObjects/UserId";
 import { AggregateRoot } from "@/contexts/Shared/domain/AgregateRoot";
