+++ src/contexts/CoreContext/domain/interfaces/services/ISkillService.ts
@@ -1,4 +1,4 @@
-import { Skill } from "../../valueObjects/Skill";
+import { Skill } from "../../entities/Skill";
 
 export interface ISkillService {
   add(skill: Skill): void;
